Not exposing internal state any more
